Move employee auth types under lib-common, unify auth status handling
