+++ frontend/src/employee-frontend/api/auth.ts
@@ -3,8 +3,8 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 import { client } from './client'
-import { AdRole, User } from '../types'
 import { JsonOf } from 'lib-common/json'
+import { AuthStatus } from 'lib-common/api-types/employee-auth'
 
 export const logoutUrl = `/api/internal/auth/saml/logout?RelayState=/employee/login`
 
@@ -16,12 +16,6 @@ const redirectUri =
 export function getLoginUrl(type: 'evaka' | 'saml' = 'saml') {
   const relayState = encodeURIComponent(redirectUri)
   return `/api/internal/auth/${type}/login?RelayState=${relayState}`
-}
-
-export interface AuthStatus {
-  loggedIn: boolean
-  user?: User
-  roles?: AdRole[]
 }
 
 export async function getAuthStatus(): Promise<AuthStatus> {
