+++ frontend/src/employee-frontend/api/unit.ts
@@ -21,7 +21,7 @@ import {
   VisitingAddress
 } from '../types/unit'
 import { ServiceNeed, UnitBackupCare } from '../types/child'
-import { AdRole, DayOfWeek } from '../types'
+import { DayOfWeek } from '../types'
 import { JsonOf } from 'lib-common/json'
 import LocalDate from 'lib-common/local-date'
 import FiniteDateRange from 'lib-common/finite-date-range'
@@ -41,6 +41,7 @@ import {
   ChildReservations,
   UnitAttendanceReservations
 } from 'lib-common/api-types/reservations'
+import { AdRole } from 'lib-common/api-types/employee-auth'
 import { UUID } from 'lib-common/types'
 
 function convertUnitJson(unit: JsonOf<Unit>): Unit {
