+++ frontend/src/employee-frontend/components/employees/EmployeePage.tsx
@@ -14,11 +14,11 @@ import { EmployeeUser } from '../../types/employee'
 import { useParams } from 'react-router-dom'
 import { SpinnerSegment } from 'lib-components/atoms/state/Spinner'
 import ErrorSegment from 'lib-components/atoms/state/ErrorSegment'
-import { GlobalRole, globalRoles } from '../../types'
 import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
 import Checkbox from 'lib-components/atoms/form/Checkbox'
 import ReturnButton from 'lib-components/atoms/buttons/ReturnButton'
 import AsyncButton from 'lib-components/atoms/buttons/AsyncButton'
+import { GlobalRole, globalRoles } from 'lib-common/api-types/employee-auth'
 
 interface FormData {
   globalRoles: GlobalRole[]
