+++ frontend/src/employee-frontend/utils/roles.tsx
@@ -4,7 +4,7 @@
 
 import React, { ReactNode, useContext } from 'react'
 import { UserContext } from '../state/user'
-import { AdRole } from '../types'
+import { AdRole } from 'lib-common/api-types/employee-auth'
 
 export const requireRole = (roles: AdRole[], ...requiresOneOf: AdRole[]) => {
   const requiredRoles: AdRole[] = [...requiresOneOf, 'ADMIN']
